+++ flake.nix
@@ -162,7 +162,7 @@
       system = "x86_64-linux";
       nixos = true;
       modules = builtins.attrNames {
-        inherit (modules) duckdns gaming i3 samba sway;
+        inherit (modules) docker duckdns gaming i3 samba sway;
       };
     }
     {